Financial/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.44.1",
]
